+++ src/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './useApp';
+export { useScrollSpy } from './useScrollSpy';
