+++ src/utils/index.ts
@@ -1,1 +1,1 @@
-export { useScrollSpy } from './useScrollSpy';
+
